refactor: move phenotypic.nn → detect.nn and rename RemoveNonCircular → RemoveLowCircularity
- `phenotypic.nn` subpackage relocated to `phenotypic.detect.nn`, grouping
  GPU-backed detectors alongside the rest of the detection layer; the old
  top-level `phenotypic.nn` import no longer exists
- `RemoveNonCircular` renamed `RemoveLowCircularity` (file, class, and all
  call-sites) to better reflect that the filter acts on the circularity
  score rather than on "non-circularity" as a property
- Updated all import paths, docstrings, prefab pipelines, tests (moved to
  tests/unit/detect/nn/), docs, and README to reflect both changes
