+++ pyproject.toml
@@ -133,6 +133,7 @@ dev = [
     "pandas-stubs>=2.3.3.260113",
     "types-tqdm>=4.67.3.20260518",
     "types-psutil>=7.2.2.20260518",
+    "pyqt6>=6.9.1",
 ]
 docs = [
     "sphinx",
