+++ src/phenotypic/__init__.py
@@ -37,7 +37,6 @@ from . import (
     enhance,
     grid,
     measure,
-    nn,
     refine,
     schema,
     settings_,
@@ -62,7 +61,6 @@ __all__ = [
     "prefab",
     "correction",
     "enhance",
-    "nn",
     "tools_",
     "util",
     "settings_",
