+++ src/phenotypic/_core/_pipeline_parts/_serializable_pipeline.py
@@ -598,7 +598,7 @@ class SerializablePipeline(NapariPipelineViewer):
             "phenotypic.analysis",
             "phenotypic.prefab",
             "phenotypic.post",
-            "phenotypic.nn",
+            "phenotypic.detect.nn",
         ]
 
         for module_name in submodules:
