+++ src/phenotypic/detect/nn/_checkpoint_manager.py
